requirements.txt
streamlit
pandas
requests
openpyxl
plotly.express